+++ packages/core/src/lib/express.ts
@@ -1,13 +1,11 @@
 import { Application, RequestHandler } from 'express';
-import { RedisClientOptions, RedisClientType } from 'redis';
 import cors from 'cors';
 import helmet from 'helmet';
 import bodyParser from 'body-parser';
-import rateLimit, { Options, Store } from 'express-rate-limit';
+import rateLimit, { Options } from 'express-rate-limit';
 import { PathParams } from 'express-serve-static-core';
 
 import * as defaults from '../config/app.config';
-const redisStore = require('rate-limit-redis');
 
 /**
  * Set some basic security measurements
@@ -47,24 +45,13 @@ export function setBodyParser(
  * Current support for: built-in memory and Redis
  */
 export function getRateLimiter(options: RateLimiterOptions = {}): RequestHandler {
-  let store: Store | undefined;
   const allOptions: any = Object.assign({}, defaults.rateLimiterOptions, options);
 
-  // Automatically assign new redis store instance
-  if (allOptions.redis) {
-    store = new redisStore(allOptions.redis);
-  }
-
-  const { redis, ...rateOptions } = allOptions; // Filter out unneeded properties
-  return rateLimit({
-    ...rateOptions,
-    store,
-  });
+  return rateLimit(allOptions);
 }
 
 // Interfaces
 export interface RateLimiterOptions extends Partial<Options> {
-  redis?: RedisOptions;
 }
 
 export interface SecurityOptions {
@@ -77,8 +64,4 @@ export interface BodyParserOptions {
   raw?: bodyParser.Options;
   text?: bodyParser.OptionsText;
   urlEncoded?: bodyParser.OptionsUrlencoded;
-}
-
-export interface RedisOptions extends RedisClientOptions {
-  client?: RedisClientType;
 }